requirements.txt
--extra-index-url=http://dist.pinaxproject.com/dev/
--extra-index-url=http://dist.pinaxproject.com/alpha/

Django==1.4.1
pinax-theme-bootstrap==2.0.4
pinax-theme-bootstrap-account==1.0b2
django-user-accounts==1.0b3
django-forms-bootstrap==2.0.3.post1
metron==1.0
pinax-utils==1.0b1.dev3
ipython
south
django-simple-captcha
PIL
django-debug-toolbar
mysql-python
simplejson
python-memcached
pycrypto
boto
django-ses
django-widget-tweaks
